avanços em controllers
