+++ package/__main__.py
@@ -1,0 +1,8 @@
+from package.controller.SistemaController import SistemaController
+
+def main():
+    app = SistemaController()
+    app.initialize()
+
+if __name__ == "__main__":
+    main()
