+++ package/controller/ClienteController.py
@@ -1,26 +1,22 @@
-from model.Singleton import Singleton
-from model.Banco import Banco
-from view.BancoView import BancoView
-from config import Config
-from exception.FaultyInputException import FaultyInputException
+from package.model.Cliente import Cliente
+from package.model.Singleton import Singleton
+from package.config import Config
+from package.exception.FaultyInputException import FaultyInputException
+from package.view.ClienteView import ClienteView
 
 
-class BancoController(Singleton):
+class ClienteController(Singleton):
     def __init__(self, controlador_sistema):
-        self._banco: Banco = Banco()
-        self._view: BancoView = BancoView()
+        self._clientes: list[Cliente] = []
+        self._view: ClienteView = ClienteView()
         self._sistema = controlador_sistema
-
-    @property
-    def banco(self) -> Banco:
-        return self._banco
 
     def initialize(self):
         options = {
             1: self._show_clientes,
-            2: self._registrar_cliente,
+            2: self._criar_cliente,
             3: self._remover_cliente,
-            4: self._retornar
+            4: self._retornar,
         }
 
         while True:
@@ -32,10 +28,9 @@ class BancoController(Singleton):
                 pass
 
     def _show_clientes(self):
-        for cliente in self._banco.clientes:
-            pass
+        pass
 
-    def _registrar_cliente(self):
+    def _criar_cliente(self):
         pass
 
     def _remover_cliente(self):
