+++ package/view/ClienteView.py
@@ -1,7 +1,7 @@
-from model.Singleton import Singleton
+from package.model.Singleton import Singleton
 
 
-class ContaView(Singleton):
+class ClienteView(Singleton):
     def __init__(self):
         pass
