--- src/view/ClienteView.py
@@ -1,10 +1,0 @@
-from model.Singleton import Singleton
-
-
-class ClienteView(Singleton):
-    def __init__(self):
-        pass
-
-    def show_message(self, message: str):
-        print(message)
-
